+++ packages/react-components/src/components/Card/index.ts
@@ -1,0 +1,9 @@
+export { Card } from './Card';
+export type {
+  CardActionLink,
+  CardDismissButton,
+  CardIcon,
+  CardMetadataItem,
+  CardProps,
+  CardTag,
+} from './Card';